add search

+++ src/routes/rating.py
@@ -1,7 +1,8 @@
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, status, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 from uuid import UUID
 from typing import List
+
 from src.database.db import get_db
 from src.repository.rating import RatingRepository
 from src.schemas.rating import RatingCreate, RatingResponse, AverageRatingResponse
@@ -32,13 +33,13 @@ async def create_rating(
 
 @router.get("/average-rating/{photo_id}", response_model=AverageRatingResponse)
 async def get_average_rating(
-    photo_id: UUID,
-    db: AsyncSession = Depends(get_db),
+		photo_id: UUID,
+		db: AsyncSession = Depends(get_db),
 ) -> AverageRatingResponse:
-    await photo_repository.get_photo_by_id_or_404(photo_id, db)
-    avg_rating = await RatingRepository.get_average_rating(db, photo_id)
-    avg_rating_rounded = round(avg_rating, 2)
-    return AverageRatingResponse(average_rating=avg_rating_rounded)
+	await photo_repository.get_photo_by_id_or_404(photo_id, db)
+	avg_rating = await RatingRepository.get_average_rating(db, photo_id)
+	avg_rating_rounded = round(avg_rating, 2)
+	return AverageRatingResponse(average_rating=avg_rating_rounded)
 
 
 @roles_required((Role.admin, Role.moderator))
@@ -47,15 +48,25 @@ async def delete_rating(
 		photo_id: UUID,
 		user_id: UUID,
 		db: AsyncSession = Depends(get_db),
+		current_user: User = Depends(UserRepository.get_current_user)
 ) -> None:
-	await RatingRepository.delete_rating(db, user_id, photo_id)
+	if not current_user.is_admin and current_user.is_moderator:
+		raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='You cannot do it')
+	else:
+		await RatingRepository.delete_rating(db, user_id, photo_id)
 	return None
 
 
+@roles_required((Role.admin, Role.moderator))
 @router.get("/{photo_id}", response_model=List[RatingResponse])
 async def get_ratings_for_photo(
 		photo_id: UUID,
-		db: AsyncSession = Depends(get_db)
+		db: AsyncSession = Depends(get_db),
+		current_user: User = Depends(UserRepository.get_current_user)
 ) -> List[RatingResponse]:
 	ratings = await RatingRepository.get_ratings_for_photo(db, photo_id)
+
+	if not current_user.is_admin and current_user.is_moderator:
+		raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='You cannot do it')
+
 	return ratings
